requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
pytest
httpx
freezegun